2025-04-17 MAC & IP addresses, & Subnet Mask.
Today’s class was all about MAC addresses, IP addresses, and how the internet works. It was a lot of new information, but very interesting!

We learned the difference between a MAC address and an IP address:

A MAC address is a permanent number (48 bits) that is kind of like your computer’s name tag. It’s written in hexadecimal.

An IP address is not permanent—it can change. It tells the internet where your device is so data can find it.

We also talked about subnet masks. These are used to break up networks into smaller parts. There are different classes of networks:
Class         Default Subnet Mask     CIDR Notation    IP Range
Class A  =  255.0.0.0                         /8                        0.0.0.0 to 127.255.255.255

Class B =  255.255.0.0                     /16                       128.0.0.0 to 191.255.255.255

Class C =  255.255.255.0                /24                       192.0.0.0 to 223.255.255.255

Class D = Unknown                           N/A                      224.0.0.0 to 239.255.255.255

Class E = Unknown                            N/A                     240.0.0.0 to 255.255.255.255

We looked at how each network has a network number, and then we assign device addresses like:

X.0.0.1

X.0.0.2

And so on, until the last one before the broadcast address (like 115.255.255.254).

We learned that the broadcast address sends data to all devices on the network, and it always ends in 255.

We also practiced counting the number of IP addresses using this chart:

128, 64, 32, 16, 8, 4, 2, 1
That helps us figure out how many devices can be in a subnet.

We also learned about:

CIDR – Classless Inter-Domain Routing. It helps break networks into smaller chunks, and is written like /25, /26, /27, and so on.

Mr. Darrel explained that if you're given an IP address and a subnet mask, you can figure out:

The network number

The first device address

The second, third, and last address

The broadcast address

Today helped me better understand how computers and the internet talk to each other. MAC and IP addresses are like names and addresses for devices. Subnet masks and CIDR help organize everything so the internet doesn’t get confused. I’m excited to keep learning how all of this connects in real life and in cybersecurity!